refactor hooks to be more modular
