+++ benchmark/results/processed/correlations.txt
@@ -1,4 +1,4 @@
-Size vs accuracy Pearson correlation: 0.2329642529704879
-Size vs accuracy Pearson correlation p-value: 2.3337587190507702e-11
-Quantisation vs accuracy Pearson correlation: 0.23317712495853615
-Quantisation vs accuracy Pearson correlation p-value: 2.2357594347011965e-11
+Size vs accuracy Pearson correlation: 0.23390145037154594
+Size vs accuracy Pearson correlation p-value: 1.2198001457703588e-11
+Quantisation vs accuracy Pearson correlation: 0.23577777304999908
+Quantisation vs accuracy Pearson correlation p-value: 8.269650238065426e-12
