+++ benchmark/results/processed/naive_vs_biochatter.txt
@@ -1,4 +1,4 @@
-mean: 0.8395405845323136 vs 0.4921221759794938
+mean: 0.8395405845323135 vs 0.4921221759794938
 std: 0.12140586543599144 vs 0.11974304260613337
-p-value: 1.026266555798005e-38
-t-statistic: 17.761473671346472
+p-value: 1.0262665557980633e-38
+t-statistic: 17.761473671346465
